travaille sur les structures de données

+++ coordinator.py
@@ -2,9 +2,14 @@ import socket
 import signal
 import os
 import time
-def sig_usr1_handler(pid):
-    time.sleep(2)
-    os.kill(pid, signal.SIGUSR2)
+from multiprocessing import shared_memory
+
+def sig_usr1_handler(pid,shm):
+    
+    print(bin(shm.buf[0]))
+    if shm.buf[0] in [0b1000,0b0100, 0b0010, 0b0001]:
+        time.sleep(2)
+        os.kill(pid, signal.SIGUSR2)
 
 import shared_memory_process
 HOST = "localhost"
@@ -15,8 +20,9 @@ if __name__ == "__main__":
     try:
         
         data_process=shared_memory_process.shared_memory_manager('coordinator')
+        existing_shm = shared_memory.SharedMemory(name="trafficLightState")
+        signal.signal(signal.SIGUSR1, lambda a, b: sig_usr1_handler(light_pid,existing_shm))
         light_pid=int([i for i in data_process if i[0]=='light'][0][1])
-        signal.signal(signal.SIGUSR1, lambda a, b: sig_usr1_handler(light_pid))
         conn = None
         server_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
         server_socket.bind((HOST, PORT))
